Make availability page mobile-friendly and hide the timezone input
The purpose of this commit was to hide the timezone input button.
This was because you could scroll on the right because the input
was styled with absolute positioning and told to stay on the right.
To remedy this, the input field was hidden as the value would be
sent in the form POST anyway. This was also a good oppurtunity to
make the page mobile-responsive. This was done by using bootstrap's
grid system. This was done because the page looked very bad on
mobile otherwise.