Add anson and bariol fonts to edit_availability.css
The reason this wasn't done before was because prior to this commit,
the fonts anson and bariol were stored in cache, so setting them as
a font on a text worked. However, for a user who doesn't have the fonts
on cache, anson and bariol fonts won't be found. However, now this
problem won't happen for these fonts because we added them manually in
the css file. However, because of this cache issue, there are going to
be more fonts that we haven't noticed that weren't explicitly created in
the css files.