Allow multiple availabilites on the same day, show times in user's timezone
Prior to this commit, you couldn't add multiple availabilties on the
same day, your new availability would just override the first
availability set on that day. Now, however, you can set multiple
availabilities on the same day. There are limitations to the current
code.

First, design wise, the limitations are that if you set multiple
availabilities on the same day, two rows will show up and the second
row will state the name of the week again. However, I don't see this
as a huge issue as I think eventually, we want to update the page to
not display availabilites the way it does right now, as I don't think
the design looks good. CodyHouse as a good design for a schedule-template,
replacing the current system to edit availability with that would be
smart I think.

Second, another limitation to the current codebase is that you can set an
availability for Monday from 8:00 pm to 10:00 pm and then subsequently,
set another availability for Monday from 8:30 pm to 9:30 pm. This issue
should be remedied as a user shouldn't be allowed to have availabilites in
their existing availabilities.

These are the limitations of our current system of editting availabilities.
I think we should start finding a template from other people to edit
availabilities ASAP.

One of the bigger successes in this commit was that users can now view other
user's availabilities in thier local timezones. It can be argued that a
feature like this isn't necessary but I think it is very necessary as I
wouldn't want the limitations of our website to restrict a user's needs from
schedulearn.

The limitations of making timings display locally is that the code in
edit_availability.html is very complex, however considering that we would
eventually want to move on from this design, I don't think the complex code
is a serious issue.