Provide error-checking for overlapping availability timings
Prior to this commit, no errors were returned if an availability timing
overlapped with another timing. This is no longer an issue. There are
limitations to this fix as the values of what they entered as timings
and day are not returned if the error is triggered, meaning the user has
to remember what availability he set. This is not good and needs to be
remedied in a future branch.