Submit availability form using ajax (edit_availability)
Purpose of this commit was to stop submitting the availability form
using html and do so using ajax. The advantage of this is that the
website doesn't need to refresh if there is an error in the input
form and this way keeps the html code cleaner. Eventually, we'd want
to remove the availability table code from the html to the js too,
but for this commit, I'll leave it at availability form.