Add feature to save lessons to google calendar as events
A lot of things were changed in this commit:

- Permission to manage calendars only comes up when someone decides to schedule
a lesson, unfortunately this means that anytime a user wants to save a lesson,
they need to pick a google account. But I think this flow is better than the
one in the previous commit.

- Add a button when a lesson has been confirmed that allows to save the lesson
in google calendars.

- Create another authentication backend specifically for saving lessons in
google calendar

- Rescheduling an icon changed from a calendar to a pencil icon.

- New backend view that manages saving a lesson to g-calendars and then
redirects to agenda after it's done.