Check for authorization, add location, status, and attendees to gcalender event
With this commit, if someone tries to save a lesson to their google calender
using a url, they won't be able to because the view checks if the request.user
is part of the lesson. It also adds location, event status and attendees to
the google event, and finally it removes a print stmnt.