Increase width of tabs, increase factor of slider movement
I noticed a design inconsistency where when you clicked on the
availability tab vs another tab, the edges of the availability
tab were right beside the text whereas the other tabs had
spacing between the end of the text and the edge of the tab
box. To remedy this issue, I increased the width so every tab
had even spacing. This raised an issue; since the width was
bigger, the slider wasn't moving far enough to underline the
next tab that was clicked on. To fix this I increased the
factor the slider would go from 100 to 120 which fixed the
slider's position